submissions/requirements.txt
# UdaPlay Project Requirements
# Python 3.11+

# Core dependencies
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0

# Additional dependencies used by the lib modules
pydantic>=2.0.0  # Used for data validation in the lib modules
pdfplumber>=0.9.0  # Required by lib/loaders.py (even though we don't use it directly)